requirements.txt
Pillow==9.5.0
jurigged==0.5.4